[MOVE] contextualexecutorパッケージを移動する
contextualexecutor -> seichiassistの依存は存在しない

なんならモジュール分けてもいいけど
依存管理が面倒くさいのでとりあえず同モジュールに残す
