+++ src/main/java/com/github/unchama/contextualexecutor/builder/TypeAliases.kt
@@ -1,12 +1,12 @@
-package com.github.unchama.seichiassist.commands.contextual.builder
+package com.github.unchama.contextualexecutor.builder
 
 import arrow.core.Either
 import arrow.core.Option
 import arrow.effects.IO
-import com.github.unchama.seichiassist.commands.contextual.ParsedArgCommandContext
-import com.github.unchama.seichiassist.commands.contextual.PartiallyParsedArgs
-import com.github.unchama.seichiassist.commands.contextual.RawCommandContext
-import com.github.unchama.seichiassist.commands.contextual.builder.response.ResponseToSender
+import com.github.unchama.contextualexecutor.ParsedArgCommandContext
+import com.github.unchama.contextualexecutor.PartiallyParsedArgs
+import com.github.unchama.contextualexecutor.RawCommandContext
+import com.github.unchama.contextualexecutor.builder.response.ResponseToSender
 
 typealias CommandResponse = Option<ResponseToSender>
