+++ src/main/java/com/github/unchama/contextualexecutor/executors/BranchedExecutor.kt
@@ -1,9 +1,9 @@
-package com.github.unchama.seichiassist.commands.contextual.executors
+package com.github.unchama.contextualexecutor.executors
 
 import arrow.effects.IO
 import arrow.effects.extensions.io.applicative.just
-import com.github.unchama.seichiassist.commands.contextual.ContextualExecutor
-import com.github.unchama.seichiassist.commands.contextual.RawCommandContext
+import com.github.unchama.contextualexecutor.ContextualExecutor
+import com.github.unchama.contextualexecutor.RawCommandContext
 
 /**
  * コマンドの枝分かれでのルーティングを静的に行う[ContextualExecutor]
